Fix for 'download' on media detail page on Honeycomb
DownloadManager.Request.VISIBILITY_VISIBLE_NOTIFY_COMPLETED is
documented as appearing in 3.0 (api 11) but causes a crash on
a GoogleTV 2 (Honeycomb-based) device.

Skipping it for Honeycomb as we do for Gingerbread; your media
goes into the 'Downloads' directory and can be found there, but
there's no notification on completion.

GitHub: https://github.com/wikimedia/apps-android-commons/pull/24
Change-Id: I4b5b0c0ffc0d9bfa8d1ff524d531022ba8f3f9f8