Crash fix for rotate in media detail view
We were crashing when rotating in the detail view, because the detail
view fragment had assumed that its parent activity had already loaded
up its cursor and such before it got created -- an assumption which
fails dramatically when rotating, or switching away and back after
we got paged out of memory.

Check whether we got an object, and if not ask the parent to ping us
when it's done loading.

Also tweaked one of the older delay hacks to avoid a second crash
related to the pager's adapter.

Change-Id: Ib91c0fef0d8ffd0e6b8e0e4874f3ee9be3e9cf34