fix for #667 issue https://github.com/commons-app/apps-android-commons/issues/667
This bug can be reproduced on low/bad connections only. Because of asynchronus work of getting new categories

Problem was in async requests, and reaction on cancel was wrong.
First of all to decrease impact on phone internet connection - request should start on after text changed, and cancel on before text changed to deliver proper results.

Also requesting on create view was unnecessary, because results may deliver BEFORE view attached to fragment