docs: add an initial version of commit message/contributors guideline
The project has long evolved without a guideline for commit messages
and a contributors guideline. This has led to a lot mysteries in the
code base.

To avoid such mysteries in the code base in the future, add an intial
version of the commit message guideline in the form of a contributors
guideline.

The name 'CONTRIBUTING' is pretty common and also ensures that GitHub
(the place where this project currnetly evolves) brings the document to
notice to an user trying to create Pull Request. Further, the all caps
ensure it catches the eyes of the gazing user.

This initial version of guideline is based on the git project's guideline
for submitting patches (Documentation/SubmittingPatches).