contributing: move the main contributing file to the wiki
The document was based on Git project's guidelines for submitting
patches (Documentation/SubmittingPatches). As it was licensed under
GPLv2 while this porject is licensed under an Apache license a license
conflict arose.

To solve it move the main document which might be susceptible to
license issues to the wiki while keeping a jist of it in the
document itself.