Add CONTRIBUTING file (#879)
* docs: add an initial version of commit message/contributors guideline

The project has long evolved without a guideline for commit messages
and a contributors guideline. This has led to a lot mysteries in the
code base.

To avoid such mysteries in the code base in the future, add an intial
version of the commit message guideline in the form of a contributors
guideline.

The name 'CONTRIBUTING' is pretty common and also ensures that GitHub
(the place where this project currnetly evolves) brings the document to
notice to an user trying to create Pull Request. Further, the all caps
ensure it catches the eyes of the gazing user.

This initial version of guideline is based on the git project's guideline
for submitting patches (Documentation/SubmittingPatches).

* contributing: remove internal link

Markdown doesn't seem to be supporting internal links(fragments) to other
parts of the document.

So, replace it with normal text to refer to the section.

* contributing: fix typo and improve sentence

* contributing: add link to related wiki page

* contributing: fix typo

* contributing: quote section name correectly

Only part of the name of a section was quoted thus referring to
an inexistent section and making the sentence meaningless.

So, quote it completely.

* contributing: use plural to refer to wiki pages

A change might obsolete "multiple" Wiki pages, which have
technical documentation about the app.

So, use plural to refer to the wiki pages.

* contributing: move the main contributing file to the wiki

The document was based on Git project's guidelines for submitting
patches (Documentation/SubmittingPatches). As it was licensed under
GPLv2 while this porject is licensed under an Apache license a license
conflict arose.

To solve it move the main document which might be susceptible to
license issues to the wiki while keeping a jist of it in the
document itself.

* contributing: use lenient wording for steps that might be optional

It's not possible to write tests for all changes. For example, UI
changes. Also, it's not the case that the Wiki pages become stale all
the time.

So, use lineant wordings for optional steps

* contributing: correct grammar in a sentence

"documented in" reads better than "documented at" (at least for
me).

If this is wrong then we can revert it.

* contributing: add link to guidelines in the wiki page

* contributing: replace 'jist' with 'gist'

  "Jist is a pronunciation spelling of the word just when spoken
   with an American Southern accent.

   Gist is a noun that means a quick summary or the main point of
   something."

I intended to use the latter. So, ..