Fix tamil translation of app name
The app name has been 'Commons' (and not 'Wikimedia Commons') for a
very long time. For completely unknown reasons, the Tamil (ta)
translation of the app name stuck with the translation for 'Wikimedia
Commons' regardless of the fact that the translation had been
corrected at translatewiki.net in the year 2016. The correction
somehow didn't reflect here during the localisation updates from
tranlsatewiki.net.

Ignoring the weirdness, correct the translation of the app name
to reflect it's current name.