Bugfix(#1921)
Title

Fixes #1921

Description

Fixes #1921

The reported issue is that in the Nearby activity, the app does not communicate that there is no internet connection.
On investigation, it appears that a toast message does briefly display "No internet connection".
It looks like other activities have implemented this same message within the snackbar instead.
In order to have a consistent style of communication to the user, I removed the toast message and replace it with the snackbar message, as implemented in NotificationActivity.java.

Tests performed

Tested on API 28 with Samsung Galaxy s8+ via USB Debugging.

1) Log into application		->	You are taken to the "My Recent Uploads" Activity.
2) Turn on airplane mode	->	Your device loses its internet connection.
3) Tap on "Nearby" 			->	The nearby activity will load  the "No Internet Connection" snackbar at the bottom.
								The snackbar message will remain on the screen.
4) Turn off airplane mode	->	Your device regains its internet connection. 
								The Nearby activity will automatically load its contents when the connection is re-established.