Fixed the infowindow duplication
The cause of the duplication was the inability to deselect the nearest place info window after the list of nearby places was updated. As such, the previous infowindow would be permanently rendered to the screen, and a new infowindow would be rendered overlapping the old one partially.

I moved the code which de-selects markers to be called before the list of nearby places is updated, and the duplication of the infowindow no longer occurs.

An easy test was to change the orientation of the device, causing the nearby map fragment to be re-created.

Also, removed the getNearestMarker function, which was added to the previous commit in error.