Use all the continue node attributes in API response (Fixed #2182)
This patch saves all the continue node attributes in API response for
following use. This fix a bug when only part of the continue
parameters are saved and used, which causing client to request the same
image indefinitely.

It also removes category continue node attributes from SharedPrefs,
since these attributes are transient and don't need to be saved
permanently.