Rephrase misleading statement shown when skipping login
The previous statement was misleading as it could also
mean that the user could never upload pictures if he
skips login. This not the case because after skipping
the login he could login again if he wants to upload
pictures.

Clarify it by using a statement which states that he
would have to login in future in case he wants to
upload pictures.