Save more properties to Contribution database(Fixed #2547)
Some useful properties of class Contribution are not saved to
Contribution database. This caused some inconvenience when re-upoading
falied images. Such that the image's date source, location and
categories are lost when re-uploading. These properties should also be
saved to database and reused when re-uploading failed images. By this
the user does not need to add/edit these properties manually.