Temporarily remove instrumented test(s).
The project's instrumented tests consist of a single test which sets a
SharedPreference, then verified that the SharedPreference was set.

I humbly submit that this is not worth the trouble of having the CI
server download the system image, spin up an emulator, load the APK, and
run it just for this purpose.

Once the instrumented tests are ready for prime time, we can reinstate
instrumented tests in CI.