CategoryImagesListFragment: fix inappropriate error shown when there are no more images to load string.xml: added appropriate message to show when no more images are left to load
When the featured or uploaded by mobile list comes to end, i.e there are no more images to load an error is shown instead of suitable information.

The recent update of Rxjava does not allow Single to return null value which is the case when there are no more images left and leads to NullPointerException.
By checking the error for empty list a suitable information could be provided to the user.

Check throwable message for null value in handleError method in CategoryImagesListFragment class and show appropriate snackbar if needed.