Rename Mapbox telemetry setting (#3995)
Fixes #3946

"Telemetry Opt Out" is a contradictory name for the telemetry setting
as it contradicts the option description by conveying that enabling
the setting means the user has opted out of telemetry while that is
not the case. Enabling the setting means telemetry would be sent
to Mapbox.

So, use a better name that conveys the meaning in a better way without
any contradiction.