#171: Bump Triple-T gradle-play-publisher
For alpha builds we use the gradle-play-publisher to automatically package and publish the app. In older versions there was a compatibility bug with Gradle 6.3 (https://github.com/Triple-T/gradle-play-publisher/issues/797). Using the updated version (2.8.1) should fix the problem.

There is version 3.0.0, however I've stuck to 2.8.1 as the breaking change of 3.0.0 is that it no longer supports PKCS keys (.p12 files) which we use.