Updating comments in PageEditClient.java
Description (required)

Fixes commons-app#4161
In the comments of the class PageEditClient.java "@param pagetitle" is written in three areas in this file. It should have been "@param pageTitle" with capital T as pageTitle is being passed as an argument rather than pagetitle.

What changes did you make and why?
Changes are made in the comments of fil where the parameters for Observable method are stated. pagetitle has been changed to pageTitle as the correct argument passed. This change would increase the code readability for fellow developers.