Reduce minSdk back to 21 (#5236)
Commit 0f0358f22 (Bump minSdk to 29, 2023-04-30) bumped the minSdk to
29 since it was found to be necessary for publishing the app to Play
Store.

On futher analysis[1], it seems like that might actually not be necessary.
The crux is:

  1. We (rightly) answer "No" to the Play Console question:

       "Does your app access location in the background in APKs or
        app bundles targeting Android 9 (Pie) or older (SDK level 28 or lower)?"

  2. We do not request for the ACCESS_BACKGROUND_LOCATION permission.

So, it should be safe for us to have the minSdkVersion as 21 itself. If
again we face any rejection of the app update, let us handle it at that
point. I clearly am not able to see any valid reason for Play store to
reject the app update.

[1]: https://github.com/commons-app/apps-android-commons/issues/5223#issuecomment-1569554999

Co-authored-by: Kaartic Sivaraam <kaartic.sn@zohocorp.com>