Switch to using the commons toolforge instance fully
So far we used both the urbanecmbot instances and also the commons-android-app
toolforge instance for the functioning of the app.

The reality is that the commons-android-app instance itself had the ability
to return all the data necessary for the app. Use the same to get all the
data necessary for the app.

Fixes #5462