Remove the value-yue-hant file
The file is not properly recognized by Android and we've actually
codemapped it to yue. The translatewiki configuration has been
done in the incorrect file. So, it is still being created.

The following Gerrit change would correct it for updates after that
change is merged.

  https://gerrit.wikimedia.org/r/c/translatewiki/+/1022508

For the time being remove it for the sake of release.