app: adjust R8 rules to ensure our model classes are not obfuscated
With refactoring of the dataclient module, the model classes now live
within the app's source code itself. So, the existing R8 rules became
obsolete and resulted in the prodRelease version of the app not
working.

So, adjust the R8 rules so that R8 doesn't obsfuscate the model classes
that now live within the app.